requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

adal==1.2.7
azure-common==1.1.27
azure-core==1.18.0
azure-identity==1.6.1
azure-keyvault-secrets==4.3.0
azure-mgmt-core==1.3.0
azure-storage-blob==12.9.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.8
idna==2.10
isodate==0.6.0
msal==1.14.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
numpy==1.20.3
oauthlib==3.1.1
pandas==1.2.4
portalocker==1.7.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.4
job-seeker==0.0.2
